+++ scripts/export-import/related/scene.js
@@ -1,3 +1,4 @@
+import {ExtractRelatedActiveTileData} from './active-tiles.js';
 import {RelatedData} from './related-data.js';
 
 /**
@@ -37,49 +38,7 @@ export function ExtractRelatedSceneData(scene) {
     relatedData.AddRelations(uuid, tokens);
   }
 
-  // Monk's Active Tile Triggers support.
-  const activeTiles = ScenePacker.GetActiveTilesData(scene.data?.tiles);
-  if (activeTiles.length) {
-    const extractActiveTileReference = (value) => {
-      if (!value) {
-        return undefined;
-      }
-      const entityParts = value.split('.');
-      if (entityParts.length < 2 || !entityParts[0] || !entityParts[1]) {
-        // Missing definition of the entity type and entity id, unable to match.
-        return undefined;
-      }
-      return `${entityParts[0]}.${entityParts[1]}`;
-    }
-    let path = 'flags.monks-active-tiles.actions';
-    for (const tile of activeTiles) {
-      const actions = tile.getFlag('monks-active-tiles', 'actions');
-      for (const action of actions) {
-        const entityReference = extractActiveTileReference(action.data?.entity?.id);
-        if (entityReference && entityReference !== `Scene.${scene.id}`) {
-          // We have a reference to an entity that isn't this current scene.
-          relatedData.AddRelation(uuid, {uuid: entityReference, path});
-        }
-
-        const itemReference = extractActiveTileReference(action.data?.item?.id);
-        if (itemReference) {
-          relatedData.AddRelation(uuid, {uuid: itemReference, path});
-        }
-
-        if (action.data?.location?.sceneId) {
-          relatedData.AddRelation(uuid, {uuid: `Scene.${action.data.location.sceneId}`, path});
-        }
-
-        if (action.data?.macroid) {
-          relatedData.AddRelation(uuid, {uuid: `Macro.${action.data.macroid}`, path});
-        }
-
-        if (action.data?.rolltableid) {
-          relatedData.AddRelation(uuid, {uuid: `RollTable.${action.data.rolltableid}`, path});
-        }
-      }
-    }
-  }
+  relatedData.AddRelatedData(ExtractRelatedActiveTileData(scene.data?.tiles, uuid));
 
   return relatedData;
 }